// src/types/Synergetic/iSynAddress.ts
import iSynLuCountry from './Lookup/iSynLuCountry';

type iSynAddress = {
  AddressID: number;
  AddressCreatedDate: Date;
  AddressCreatedBy: string;
  AddressModifiedDate: Date;
  AddressModifiedBy: string;
  Address1: string;
  Address2: string;
  Address3: string;
  Suburb: string;
  State: string;
  CountryCode: string;
  PostCode: string;
  PhoneActual: string;
  PhoneAlt: string;
  Fax: string;
  HomeAddressSameFlag: boolean;
  HomeAddress1: string;
  HomeAddress2: string;
  HomeAddress3: string;
  HomeSuburb: string;
  HomeState: string;
  HomeCountryCode: string;
  HomePostCode: string;
  ValidAddressFlag: boolean;
  SilentPhoneFlag: boolean;
  MailFlag: boolean;
  MagazineFlag: boolean;
  SolicitationFlag: boolean;
  RaffleFlag: boolean;
  ReasonFlagChange: string;
  ReasonFlagChangeDate: Date | null;
  AddressComma: string;
  AddressFull: string;
  HomeAddressComma: string;
  HomeAddressFull: string;
  Phone: string;
  AddressRegion: string;
  AddressRegionOverrideFlag: boolean;
  AddressComment: string;
  AddressDPID: string;
  AddressSortPlanNumber: string;
  HomeAddressDPID: string;
  HomeAddressSortPlanNumber: string;
  AddressPAFVersion: string;
  HomeAddressPAFVersion: string;
  AddressDPIDBarcode: string;
  HomeAddressDPIDBarcode: string;
  SilentHomeAddressFlag: boolean;
  MailFlagChangeReason: string;
  MagazineFlagChangeReason: string;
  SolicitationFlagChangeReason: string;
  RaffleFlagChangeReason: string;
  UserFlag1: boolean;
  UserFlag2: boolean;
  UserFlag3: boolean;
  UserFlag4: boolean;
  UserFlag5: boolean;
  HomeAddressLatitude: number | null;
  HomeAddressLongitude: number | null;
  ExternalValidationFlag: boolean;
  HomeExternalValidationFlag: boolean;
  HomeAddressValidationSeq: number | null;
  AddressValidationSeq: number | null;
  Country?: iSynLuCountry;
  HomeCountry?: iSynLuCountry;
}

export default iSynAddress